+++ src/components/mod.rs
@@ -8,6 +8,8 @@ mod i18n_text;
 mod i18n_text_2d;
 #[cfg(feature = "rich_text3d")]
 mod i18n_text_3d_segment;
+#[cfg(feature = "fontmesh")]
+mod i18n_text_mesh;
 mod i18n_text_span;
 pub(crate) mod utils;
 
@@ -18,6 +20,8 @@ pub use i18n_text::*;
 pub use i18n_text_2d::*;
 #[cfg(feature = "rich_text3d")]
 pub use i18n_text_3d_segment::*;
+#[cfg(feature = "fontmesh")]
+pub use i18n_text_mesh::*;
 pub use i18n_text_span::*;
 
 /// A text component `bevy_simple_i18n` can write translated strings into.